Conflicts:
	src/mrgw.sh
	src/rgw/rgw_common.h
	src/rgw/rgw_json_enc.cc
	src/rgw/rgw_op.cc
	src/rgw/rgw_op.h
	src/rgw/rgw_rados.cc

+++ src/os/KeyValueStore.cc
@@ -649,7 +649,7 @@ int KeyValueStore::mkfs()
       goto close_fsid_fd;
     }
     if (::fsync(fsid_fd) < 0) {
-      ret = errno;
+      ret = -errno;
       derr << "mkfs: close failed: can't write fsid: "
            << cpp_strerror(ret) << dendl;
       goto close_fsid_fd;
@@ -717,6 +717,10 @@ int KeyValueStore::mkfs()
     dout(1) << g_conf->keyvaluestore_backend << " backend exists/created" << dendl;
     delete store;
   }
+
+  ret = write_meta("type", "keyvaluestore");
+  if (ret < 0)
+    goto close_fsid_fd;
 
   dout(1) << "mkfs done in " << basedir << dendl;
   ret = 0;
@@ -1189,8 +1193,10 @@ void KeyValueStore::_finish_op(OpSequencer *osr)
   if (o->onreadable_sync) {
     o->onreadable_sync->complete(0);
   }
-  op_finisher.queue(o->onreadable);
-  op_finisher.queue(to_queue);
+  if (o->onreadable)
+    op_finisher.queue(o->onreadable);
+  if (!to_queue.empty())
+    op_finisher.queue(to_queue);
   delete o;
 }
